requirements.txt
streamlit~=1.41.1
torch~=2.5.1
torchvision~=0.20.1
Pillow

matplotlib~=3.10.0